requirements.txt
asgiref==3.7.2
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==10.0.0
psycopg2-binary==2.9.7
PyJWT==2.8.0
pytz==2023.3
sqlparse==0.4.4
